# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DOCUMENTACIÓN DE APIs DEL OBSERVATORIO JOVEN",
        "description": "API de los diferentes módulos",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/sobre-cmpj": {
            "get": {
                "tags": [
                    "sobre-cmpj"
                ],
                "summary": "Listado de los sobre-cmpj",
                "description": "Devuelve un listado de los sobre-cmpj que están registrados en nuestro servidor",
                "operationId": "sobre-cmpj",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Campo por el cual ordenar los resultados. Agregar un signo '-' al principio para orden descendente (ej. -name).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de los sobre-cmpj obtenido exitosamente."
                    },
                    "401": {
                        "description": "No autorizado. Se requiere un token válido en el encabezado."
                    },
                    "403": {
                        "description": "Token no provisto. Se requiere un token en el encabezado."
                    },
                    "500": {
                        "description": "Error interno del servidor. Algo salió mal."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/sobre-cmpj/{id}": {
            "get": {
                "tags": [
                    "sobre-cmpj"
                ],
                "summary": "Obtener un sobre-cmpj por su ID",
                "description": "Devuelve un sobre-cmpj específico por su ID.",
                "operationId": "get_sobre-cmpj_by_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sobre-cmpj a obtener.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sobre-cmpj obtenido exitosamente."
                    },
                    "401": {
                        "description": "No autorizado. Se requiere un token válido en el encabezado."
                    },
                    "403": {
                        "description": "Token no provisto. Se requiere un token en el encabezado."
                    },
                    "404": {
                        "description": "sobre-cmpj no encontrado."
                    },
                    "500": {
                        "description": "Error interno del servidor. Algo salió mal."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Listado de las Categorías",
                "description": "Devuelve un listado de las Categorías que están registradas en nuestro servidor",
                "operationId": "categories",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Número de elementos por página en la paginación.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Campo por el cual ordenar los resultados. Agregar un signo '-' al principio para orden descendente (ej. -name).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Filtrar los resultados por el nombre de la categoría.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[description]",
                        "in": "query",
                        "description": "Filtrar los resultados por la descripción de la categoría.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por un término en el nombre o la descripción de la categoría.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "included",
                        "in": "query",
                        "description": "Relaciones que se deben incluir en la respuesta (posts). Separadas por comas.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de las Categorías obtenido exitosamente."
                    },
                    "401": {
                        "description": "No autorizado. Se requiere un token válido en el encabezado."
                    },
                    "403": {
                        "description": "Token no provisto. Se requiere un token en el encabezado."
                    },
                    "500": {
                        "description": "Error interno del servidor. Algo salió mal."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}